requirements.txt
Django==5.1.3
djangorestframework==3.15.2
sqlparse==0.5.2
psycopg2-binary==2.9.10
celery==5.5.0
redis
django-celery-beat